punto para poner relativo el path
Alright, so this is for those who, like me, have struggled with this issue, and have tried to find a good answer. Here it is (it's a record for me as well to remember):

First, make all of the image paths relative, by putting a "." in front of it. For example:

<img src={"./images/logo.svg"}/>
This particular image lived in the 'public/images/ folder of my project.